pyproject.toml
[tool.poetry]
name = "diploma-15"
version = "0.1.0"
description = ""
authors = ["Daniil Karlov <dskarlov@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"
django-request-logging = "^0.7.5"
pydantic = ">=1.10,<2.0"
requests = "^2.31.0"
pytest-factoryboy = "^2.5.1"
pytest = "^7.4.0"
pytest-django = "^4.5.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
ipython = "^8.14.0"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
skip-magic-trailing-coma = true
include = '\.pyi?&'
exclude = '''
/(
    \.git
    | \.venv
    |magrations
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 120
combine_as_imports = true
combine_star = true
skip_glob = [
    '**/migrations/**',
    '.git',
    '.venv'
]
